Crafted some unit tests to ensure accuracy
Now that I have the basic design plan, the testCipher.py file has been
updated with specially crafted unit tests using the `unittest` framework
to test the accuracy and behavior of the actual caesar cipher algorithm
and the functions that employ this algorithm.

This is going to be a great tool to ensure my program is developed
according to the design plan I have, as well as ensure the algorithm is
accurate when it gets implemented. Not all program testing will be able
to be done with these unit tests, but it will be a good guide as I
implement the program!