Fixed some bugs in the unit tests I created
The unit tests I created accidentally referred to 'cipherCharacter' as
just 'cipherChar', and this needed to be adjusted in the test file. In
addition, I had a small 'iterable' error in the runTests.py file that
was fixed by just adding a ',' character to make the tuple *actually* a
tuple, like I intended it to be.